fix: 회원가입 플로우 에러 해결
세션에서 불러오는 값을 서버측에선 받지 못해서 생기는 에러 해결
