+++ app/auth/signup/_components/SignupFormProvider.tsx
@@ -15,7 +15,7 @@ const SignUpFormProvider = ({ children }: { children: React.ReactNode }) => {
   const { handleSubmit } = form;
 
   const onSubmit = (formValues: SignupInputsValues) => {
-    // console.log(formValues);
+    console.log(formValues);
     // alert(JSON.stringify(formValues));
 
     // 회원가입 성공 후 입력정보 파기
